Update README.md
In this study, I aimed to learn the accuracy with which I could predict whether the Food Borne Diseases stated in the data sets causes significant illness, at the same time enhance the security aspect while dealing with EHR. In this paper I proposed a predictive sequential model with considerable functional accuracy along with a system which uses a different secret keyboard mechanism with an aim to prevent Shoulder Surfing attacks. Βy using this method of authentication, users can easily log into their systems without worrying about the Shoulder Surfers as the character which is displayed on the screen is not the same as the character which is being stored in the back end. It is a simple and effective solution even for camera based attacks. In future, this system can be used for applications which require moderate to high security, since the other conventional password schemes are vulnerable to shoulder surfing. Overall, my main goal for this research is to come up with a solution which uses existing cipher to solve modern problems.